Add error handling with longjmp for contract violations
- Introduced ERRORJUMP preprocessor directive to enable error handling via longjmp.
- Modified the CONTRACT_VIOLATED macro to use longjmp for error signaling.
- Updated various functions across the codebase to utilize the new error handling mechanism, including:
  - Internal::remove_garbage_binaries
  - External::check_assignment
  - External::check_assumptions_satisfied
  - External::check_constraint_satisfied
  - Fatal error handling in message handling functions.
- Enhanced the Options and Solver classes to support jmp_buf for error handling.
- Adjusted the FATAL macro to accommodate error codes for better debugging.
- Ensured that all relevant checks now trigger longjmp on contract violations, improving robustness.

+++ cadical-rel-2.1.3/external/satgalaxy_cadical.h
@@ -1,6 +1,7 @@
 #ifndef SATGALAXY_CADICAL_H
 #define SATGALAXY_CADICAL_H
 #include <stdint.h>
+#include <stdlib.h>
 //=================================================================================================
 
 #ifdef __cplusplus
@@ -14,7 +15,7 @@ int cadical_vars (CaDiCaLSolver *s);
 const char *cadical_error_message(int code);
 int cadical_error(CaDiCaLSolver *solver);
 void cadical_add_clause (CaDiCaLSolver *solver, const int ps[],
-                         unsigned long length);
+                         size_t length);
 void cadical_add_empty_clause (CaDiCaLSolver *solver);
 // This function can be used to check if the formula is already
 // inconsistent (contains the empty clause or was proven to be
